Divide main state and wrapper
